# add GetOutputElementElimination for nvgpu backend