Before:
```
============================================================================
---- Processing '../ngraph-models/models/tensorflow/training/resnet50/tf_function_ngraph_cluster_resnet50_training.json'
============================================================================

---- Benchmark ----
compile time: 13,361ms
692.3ms per iteration

---- Aggregate times per op type ----
ConvolutionBackpropFilters     149,720us
ConvolutionBackpropData         93,336us
Convolution                     81,421us
BatchNormTrainingBackprop       69,952us
GetOutputElement                51,534us
Add                             42,910us
BatchNormTraining               41,926us
ReluBackprop                    33,456us
Relu                            23,300us
MaxPoolBackprop                  6,268us
MaxPool                          4,175us
Reshape                          3,437us
Multiply                         2,898us
Concat                           2,335us
Sum                              2,266us
Broadcast                        1,681us
Result                           1,435us
Pad                                865us
Divide                             510us
Subtract                           383us
Dot                                263us
Max                                 45us
OneHot                              36us
LessEq                              16us
Greater                             13us
Convert                              9us
And                                  9us
Log                                  7us
Exp                                  6us
```
After:
```
============================================================================
---- Processing '../ngraph-models/models/tensorflow/training/resnet50/tf_function_ngraph_cluster_resnet50_training.json'
============================================================================

---- Benchmark ----
compile time: 12,858ms
639.2ms per iteration

---- Aggregate times per op type ----
ConvolutionBackpropFilters     150,608us
ConvolutionBackpropData         93,653us
Convolution                     81,795us
BatchNormTrainingBackprop       69,946us
Add                             42,970us
BatchNormTraining               41,927us
ReluBackprop                    33,487us
Relu                            23,257us
MaxPoolBackprop                  6,265us
MaxPool                          4,171us
Reshape                          3,446us
Multiply                         2,907us
Concat                           2,328us
Sum                              2,246us
Broadcast                        1,679us
Pad                                867us
Divide                             512us
Subtract                           395us
Result                             365us
Dot                                267us
Max                                 46us
OneHot                              31us
LessEq                              15us
Greater                             13us
Convert                              9us
And                                  9us
Log                                  6us
Exp                                  5us
```